+++ swift/Package.swift
@@ -37,6 +37,7 @@ let package = Package(
             ],
             resources: [
                 .copy("Resources/silero_vad.mlpackage"),
+                .copy("Resources/Model"),
             ]
         ),
         .executableTarget(